refactor: move `ConditionType` to condition.go.
